refactor(twig): renamed pattern files into component.yml
