+++ packages/twig-new/src/components/tabs/tabs.component.yml
